requirements.txt
Flask==2.2.2
Flask_SocketIO==5.3.3
google_auth_oauthlib==1.0.0
pip==23.0.1
protobuf==4.22.1
pyotp==2.8.0
pyserial==3.5
python-dotenv==1.0.0
requests==2.28.1
